requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Assets==0.10
Flask-Babel==0.9
Flask-Cors==1.10.2
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.1
Flask-SQLAlchemy==2.0
Flask-Security==1.7.4
Flask-Testing==0.4.2
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
WTForms==2.0.1
Werkzeug==0.9.6
aniso8601==0.92
blinker==1.3
factory-boy==2.4.1
flask-marshmallow==0.4.0
itsdangerous==0.24
marshmallow==1.2.1
nose==1.3.4
passlib==1.6.2
python-dateutil==2.3
pytz==2014.10
requests==2.5.1
six==1.8.0
speaklater==1.3
webassets==0.10.1
